requirements.txt
APScheduler==3.6.3
attrs==22.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
black==23.1.0
bs4==0.0.1
cachetools==4.2.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
idna==3.3
libcst==0.4.9
moreorless==0.4.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
python-dateutil==2.8.2
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.0
six==1.16.0
soupsieve==2.3.2.post1
stdlibs==2022.10.9
toml==0.10.2
tomli==2.0.1
tornado==6.1
trailrunner==1.2.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
usort==1.0.5
